Add Extention method for Seeding Entities & Configurations
